+++ CHANGES.rst
@@ -5,8 +5,11 @@
 Unreleased
 ==========
 
+- Added a new metric under ``/metrics`` endpoint which exposes statistics of
+  each thread pool.
+
 - Added a new metric under ``/metrics`` endpoint which exposes the number of
-  open and total connections per protocol
+  open and total connections per protocol.
 
 2018/04/16 0.2.0
 ================